Move Requester to public interface
