+++ src/main/java/org/kohsuke/github/GHBlobBuilder.java
@@ -14,7 +14,7 @@ public class GHBlobBuilder {
 
     GHBlobBuilder(GHRepository repo) {
         this.repo = repo;
-        req = new Requester(repo.root);
+        req = repo.root.createRequester();
     }
 
     /**
