+++ src/main/java/org/kohsuke/github/GHCommit.java
@@ -340,8 +340,8 @@ public class GHCommit {
      * I'm not sure how path/line/position parameters interact with each other.
      */
     public GHCommitComment createComment(String body, String path, Integer line, Integer position) throws IOException {
-        GHCommitComment r = new Requester(owner.root)
-                .with("body",body)
+        GHCommitComment r = owner.root.createRequester()
+            .with("body",body)
                 .with("path",path)
                 .with("line",line)
                 .with("position",position)
