+++ src/main/java/org/kohsuke/github/GHCommitBuilder.java
@@ -35,7 +35,7 @@ public class GHCommitBuilder {
 
     GHCommitBuilder(GHRepository repo) {
         this.repo = repo;
-        req = new Requester(repo.root);
+        req = repo.root.createRequester();
     }
 
     /**
