+++ src/main/java/org/kohsuke/github/GHDeploymentBuilder.java
@@ -10,7 +10,7 @@ public class GHDeploymentBuilder {
 
     public GHDeploymentBuilder(GHRepository repo) {
         this.repo = repo;
-        this.builder = new Requester(repo.root);
+        this.builder = repo.root.createRequester();
     }
 
     public GHDeploymentBuilder(GHRepository repo, String ref) {
