+++ src/main/java/org/kohsuke/github/GHDeploymentStatusBuilder.java
@@ -24,7 +24,7 @@ public class GHDeploymentStatusBuilder {
     /*package*/ GHDeploymentStatusBuilder(GHRepository repo, long deploymentId, GHDeploymentState state) {
         this.repo = repo;
         this.deploymentId = deploymentId;
-        this.builder = new Requester(repo.root);
+        this.builder = repo.root.createRequester();
         this.builder.with("state",state);
     }
