+++ src/main/java/org/kohsuke/github/GHGistUpdater.java
@@ -16,7 +16,7 @@ public class GHGistUpdater {
 
     GHGistUpdater(GHGist base) {
         this.base = base;
-        this.builder = new Requester(base.root);
+        this.builder = base.root.createRequester();
 
         files = new LinkedHashMap<>();
     }
