+++ src/main/java/org/kohsuke/github/GHIssueBuilder.java
@@ -15,7 +15,7 @@ public class GHIssueBuilder {
 
     GHIssueBuilder(GHRepository repo, String title) {
         this.repo = repo;
-        this.builder = new Requester(repo.root);
+        this.builder = repo.root.createRequester();
         builder.with("title",title);
     }
