+++ src/main/java/org/kohsuke/github/GHProjectCard.java
@@ -110,7 +110,7 @@ public class GHProjectCard extends GHObject {
 	}
 
 	private void edit(String key, Object value) throws IOException {
-		new Requester(root).withPreview(INERTIA)._with(key, value).method("PATCH").to(getApiRoute());
+		root.createRequester().withPreview(INERTIA)._with(key, value).method("PATCH").to(getApiRoute());
 	}
 
 	protected String getApiRoute() {
@@ -118,6 +118,6 @@ public class GHProjectCard extends GHObject {
 	}
 
 	public void delete() throws IOException {
-		new Requester(root).withPreview(INERTIA).method("DELETE").to(getApiRoute());
+		root.createRequester().withPreview(INERTIA).method("DELETE").to(getApiRoute());
 	}
 }
