+++ src/main/java/org/kohsuke/github/GHProjectColumn.java
@@ -60,7 +60,7 @@ public class GHProjectColumn extends GHObject {
 	}
 
 	private void edit(String key, Object value) throws IOException {
-		new Requester(root).withPreview(INERTIA)._with(key, value).method("PATCH").to(getApiRoute());
+		root.createRequester().withPreview(INERTIA)._with(key, value).method("PATCH").to(getApiRoute());
 	}
 
 	protected String getApiRoute() {
@@ -68,7 +68,7 @@ public class GHProjectColumn extends GHObject {
 	}
 
 	public void delete() throws IOException {
-		new Requester(root).withPreview(INERTIA).method("DELETE").to(getApiRoute());
+		root.createRequester().withPreview(INERTIA).method("DELETE").to(getApiRoute());
 	}
 
 	public PagedIterable<GHProjectCard> listCards() throws IOException {
