+++ src/main/java/org/kohsuke/github/GHRef.java
@@ -56,7 +56,7 @@ public class GHRef {
      *      Whether or not to force this ref update.
      */
     public void updateTo(String sha, Boolean force) throws IOException {
-      new Requester(root)
+      root.createRequester()
           .with("sha", sha).with("force", force).method("PATCH").to(url, GHRef.class).wrap(root);
     }
 
@@ -64,7 +64,7 @@ public class GHRef {
      * Deletes this ref from the repository using the GitHub API.
      */
     public void delete() throws IOException {
-      new Requester(root).method("DELETE").to(url);
+      root.createRequester().method("DELETE").to(url);
     }
 
     /*package*/ GHRef wrap(GitHub root) {
