+++ src/main/java/org/kohsuke/github/GHReleaseUpdater.java
@@ -14,7 +14,7 @@ public class GHReleaseUpdater {
 
     GHReleaseUpdater(GHRelease base) {
         this.base = base;
-        this.builder = new Requester(base.root);
+        this.builder = base.root.createRequester();
     }
 
     public GHReleaseUpdater tag(String tag) {
