+++ src/main/java/org/kohsuke/github/GHSubscription.java
@@ -49,7 +49,7 @@ public class GHSubscription {
      * Removes this subscription.
      */
     public void delete() throws IOException {
-        new Requester(root).method("DELETE").to(repo.getApiTailUrl("subscription"));
+        root.createRequester().method("DELETE").to(repo.getApiTailUrl("subscription"));
     }
 
     GHSubscription wrapUp(GHRepository repo) {
