+++ src/main/java/org/kohsuke/github/GHTreeBuilder.java
@@ -33,7 +33,7 @@ public class GHTreeBuilder {
 
     GHTreeBuilder(GHRepository repo) {
         this.repo = repo;
-        req = new Requester(repo.root);
+        req = repo.root.createRequester();
     }
 
     /**
