+++ src/main/java/org/kohsuke/github/GHUser.java
@@ -43,14 +43,14 @@ public class GHUser extends GHPerson {
      * Follow this user.
      */
     public void follow() throws IOException {
-        new Requester(root).method("PUT").to("/user/following/" + login);
+        root.createRequester().method("PUT").to("/user/following/" + login);
     }
 
     /**
      * Unfollow this user.
      */
     public void unfollow() throws IOException {
-        new Requester(root).method("DELETE").to("/user/following/" + login);
+        root.createRequester().method("DELETE").to("/user/following/" + login);
     }
 
     /**
